@@ -49,29 +49,31 @@
   "dependencies": {
     "cli-table": "^0.3.1",
     "commander": "^2.14.1",
-    "deasync": "^0.1.12",
+    "deasync": "^0.1.14",
     "inquirer": "^3.0.6",
-    "js-yaml": "^3.9.1",
-    "lodash": "^4.17.5",
+    "js-yaml": "^3.12.0",
+    "lodash": "^4.17.11",
     "merapi": "^0.17.1",
     "merapi-proxy": "^0.1.3",
     "superagent": "^3.8.2",
     "uuid": "^3.2.1",
-    "zaun": "^1.0.0"
+    "zaun": "^2.0.1"
   },
   "devDependencies": {
+    "@types/inquirer": "0.0.43",
     "@types/js-yaml": "^3.9.0",
     "@types/lodash": "^4.14.104",
     "@types/node": "^7.0.55",
     "@types/sinon": "^2.3.3",
-    "@types/uuid": "^3.4.3",
-    "codacy-coverage": "^2.0.2",
+    "@types/uuid": "^3.4.4",
+    "codacy-coverage": "^2.1.1",
     "mocha": "^3.2.0",
     "mocha-lcov-reporter": "^1.3.0",
-    "mocha-typescript": "^1.1.7",
-    "nyc": "^11.5.0",
+    "mocha-typescript": "^1.1.17",
+    "nyc": "^11.9.0",
     "sinon": "^3.2.0",
     "ts-node": "^3.3.0",
-    "typescript": "^2.7.2"
+    "tslint": "^5.11.0",
+    "typescript": "^3.1.6"
   }
 }
