@@ -11,11 +11,12 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const merapi_1 = require("merapi");
 const uuid_1 = require("uuid");
 const colors = require("colors");
-const inquirer = require("inquirer");
 const repl = require("repl");
 const util = require("util");
-const deasync = require("deasync");
 const Table = require("cli-table");
+const os = require("os");
+const fs = require("fs");
+const deasync = require("deasync");
 class Bot extends merapi_1.Component {
     constructor(compile, helper, tester, api) {
         super();
@@ -24,23 +25,11 @@ class Bot extends merapi_1.Component {
         this.tester = tester;
         this.api = api;
     }
-    init(name, version, options) {
-        if (version) {
-            const versionRegex = /^\d+\.\d+\.\d+$/g;
-            if (!versionRegex.test(version)) {
-                console.log("Invalid version string");
-                console.log("Command kata init <botId> <botName> [version] is deprecated, use kata init <botName> [version] instead");
-                process.exit(0);
-            }
-        }
-        if (!version) {
-            version = "0.0.1";
-        }
+    init(name, options) {
         const botDesc = {
             schema: "kata.ai/schema/kata-ml/1.0",
             name,
             desc: "My First Bot",
-            version: version || "0.0.1",
             flows: {
                 hello: {
                     fallback: true,
@@ -90,24 +79,16 @@ class Bot extends merapi_1.Component {
         this.helper.dumpYaml("./bot.yml", botDesc);
         console.log(`Initialized ${name} successfully`);
     }
-    versions(options) {
+    revisions(options) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
-                const botId = this.helper.getBotId();
-                if (!botId) {
-                    throw new Error("BOT ID HAS NOT DEFINED");
-                }
-                const { data, response } = yield this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdVersionsGet, botId);
-                if (data.versions) {
-                    const history = data.versions;
-                    console.log("Bot Versions : ");
-                    history.data.forEach((hist) => {
-                        if (hist.tag) {
-                            console.log(`- ${hist.version} (${hist.tag})`);
-                        }
-                        else {
-                            console.log(`- ${hist.version}`);
-                        }
+                const projectId = this.getProject();
+                const { data, response } = yield this.helper.toPromise(this.api.botApi, this.api.botApi.projectsProjectIdBotRevisionsGet, projectId);
+                const result = data;
+                if (result && result.data) {
+                    console.log("Bot Revision : ");
+                    result.data.forEach((hist) => {
+                        console.log(`- ${hist.revision}`);
                     });
                 }
                 else {
@@ -207,31 +188,9 @@ class Bot extends merapi_1.Component {
             }
         });
     }
-    update(options) {
+    push(options) {
         return __awaiter(this, void 0, void 0, function* () {
             const desc = this.helper.loadYaml("./bot.yml");
-            const oldVersion = desc.version;
-            let [major, minor, patch] = desc.version.split(".").map((val) => parseInt(val));
-            switch (options.rev) {
-                case "major":
-                    ++major;
-                    minor = 0;
-                    patch = 0;
-                    break;
-                case "minor":
-                    ++minor;
-                    patch = 0;
-                    break;
-                case "patch":
-                    ++patch;
-                    break;
-            }
-            if (major === undefined || minor === undefined || patch === undefined) {
-                major = major || 0;
-                minor = minor || 0;
-                patch = patch || 0;
-            }
-            desc.version = `${major}.${minor}.${patch}`;
             desc.tag = options.tag || null;
             let bot = merapi_1.Config.create(desc, { left: "${", right: "}" });
             bot = this.compile.execDirectives(bot, process.cwd());
@@ -242,29 +201,14 @@ class Bot extends merapi_1.Component {
                 yield this.updateDraft(botDesc, desc);
                 return;
             }
-            if (!botDesc.id) {
-                const id = uuid_1.v4();
-                botDesc.id = id;
-                desc.id = id;
-                try {
-                    const result = yield this.helper.toPromise(this.api.botApi, this.api.botApi.botsPost, botDesc);
-                    console.log(`BOT CREATED SUCCESSFULLY WITH VERSION ${desc.version}`);
-                }
-                catch (e) {
-                    desc.version = oldVersion;
-                    console.log(this.helper.wrapError(e));
-                }
+            try {
+                const projectId = this.getProject();
+                botDesc.id = projectId;
+                const result = yield this.helper.toPromise(this.api.botApi, this.api.botApi.projectsProjectIdBotRevisionsPost, projectId, botDesc);
+                console.log(`Push Bot Success. Revision : ${result.data.revision.substring(0, 7)}`);
             }
-            else {
-                try {
-                    const result = yield this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdPut, botDesc.id, botDesc, {});
-                    desc.version = result.data.version;
-                    console.log(`UPDATED BOT SUCCESSFULLY WITH VERSION ${desc.version}`);
-                }
-                catch (e) {
-                    desc.version = oldVersion;
-                    console.log(this.helper.wrapError(e));
-                }
+            catch (e) {
+                console.log(this.helper.wrapError(e));
             }
             this.helper.dumpYaml("./bot.yml", desc);
         });
@@ -335,17 +279,16 @@ class Bot extends merapi_1.Component {
         });
     }
     console(options) {
-        let currentSession = (options.session ? options.session : uuid_1.v4());
+        let projectId;
         let botDesc;
         try {
+            projectId = this.getProject();
             botDesc = this.helper.loadYaml("./bot.yml");
         }
         catch (error) {
             console.log(this.helper.wrapError(error));
             return;
         }
-        const botId = botDesc.id;
-        const defaultDeploymentId = "f223c9e0-6ba1-434d-8313-a9f18ca364bd";
         const con = repl.start({
             prompt: botDesc.name + ">",
             writer(obj) {
@@ -353,103 +296,60 @@ class Bot extends merapi_1.Component {
             }
         });
         con.context.text = function text(str) {
+            let currentSession = this.getLocalSession();
             const message = {
                 type: "text",
                 content: str
             };
             const body = {
-                sessionId: currentSession,
+                session: currentSession,
                 message
             };
             try {
-                const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdConversePost, botId, body));
-                return data;
+                return this.converse(projectId, body);
             }
             catch (e) {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
         con.context.button = function button(op, obj = {}) {
+            let currentSession = this.getLocalSession();
             obj.op = op;
             const message = {
                 type: "data",
                 payload: obj
             };
             const body = {
-                sessionId: currentSession,
+                session: currentSession,
                 message
             };
             try {
-                const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdConversePost, botId, body));
-                return data;
+                return this.converse(projectId, body);
             }
             catch (e) {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
-        con.context.command = function button(command, obj = {}) {
+        con.context.command = function command(command, obj = {}) {
+            let currentSession = this.getLocalSession();
             const message = {
                 type: "command",
                 content: command,
                 payload: obj
             };
             const body = {
-                sessionId: currentSession,
+                session: currentSession,
                 message
             };
             try {
-                const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdConversePost, botId, body));
-                return data;
+                return this.converse(projectId, body);
             }
             catch (e) {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
-        con.context.current = function (session) {
-            if (arguments.length) {
-                currentSession = session;
-            }
-            else {
-                return currentSession;
-            }
-        }.bind(this);
-        con.context.session = function session(name, update) {
-            try {
-                if (!arguments.length) {
-                    const res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, currentSession, "get"));
-                    return res.data;
-                }
-                else if (arguments.length === 1) {
-                    const res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, name, "get"));
-                    return res.data;
-                }
-                else {
-                    let res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, currentSession, "getOrCreate"));
-                    const session = res.data;
-                    res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdPut, botId, defaultDeploymentId, session.id, update));
-                    return res.data;
-                }
-            }
-            catch (e) {
-                return this.helper.wrapError(e);
-            }
-        }.bind(this);
-        con.context.clear = function clear(name) {
-            name = name || currentSession;
-            try {
-                const { data } = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, name, "get"));
-                const session = Object.assign({}, data);
-                if (session) {
-                    this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdDelete, botId, defaultDeploymentId, session.id));
-                }
-            }
-            catch (e) {
-                if (e.status !== 400) {
-                    return;
-                }
-                return this.helper.wrapError(e);
-            }
-        }.bind(this);
+        con.context.current = () => this.getLocalSession();
+        con.context.clear = () => this.resetSession();
         con.context.clearCaches = function clearCaches(num = 20) {
             try {
                 for (let i = 0; i < num; i++) {
@@ -460,6 +360,27 @@ class Bot extends merapi_1.Component {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
+    }
+    sync(promise) {
+        if (promise && typeof promise.then === "function") {
+            let done = false;
+            let error = null;
+            let result;
+            promise.then((res) => {
+                done = true;
+                result = res;
+            }).catch((e) => {
+                error = e;
+            });
+            deasync.loopWhile(() => {
+                return !done && !error;
+            });
+            if (error) {
+                throw error;
+            }
+            return result;
+        }
+        throw new Error("Sync only accept promises");
     }
     pull(name, version, options) {
         return __awaiter(this, void 0, void 0, function* () {
@@ -499,26 +420,61 @@ class Bot extends merapi_1.Component {
             }
         });
     }
-    sync(promise) {
-        if (promise && typeof promise.then === "function") {
-            let done = false;
-            let error = null;
-            let result;
-            promise.then((res) => {
-                done = true;
-                result = res;
-            }).catch((e) => {
-                error = e;
-            });
-            deasync.loopWhile(() => {
-                return !done && !error;
-            });
-            if (error) {
-                throw error;
-            }
-            return result;
+    getProject() {
+        const projectId = this.helper.getProp("projectId");
+        if (!projectId || projectId === "") {
+            throw Error("Error : You must specify a Project first, execute kata list-project to list your projects.");
         }
-        throw new Error("Sync only accept promises");
+        return projectId;
+    }
+    converse(projectId, body) {
+        const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.projectsProjectIdBotConversePost, projectId, body));
+        const { session } = data;
+        this.setLocalSession(session);
+        return data;
+    }
+    setLocalSession(session) {
+        const jsonPath = `${os.homedir()}/.katasession`;
+        try {
+            if (session) {
+                fs.writeFileSync(jsonPath, JSON.stringify(session), "utf8");
+            }
+        }
+        catch (error) {
+            console.log(this.helper.wrapError(`Error set local session : ${error.message}`));
+        }
+    }
+    getLocalSession() {
+        const jsonPath = `${os.homedir()}/.katasession`;
+        if (fs.existsSync(jsonPath)) {
+            return JSON.parse(fs.readFileSync(jsonPath, "utf8"));
+        }
+        else {
+            // default session
+            return {
+                "channel_id": "console-channel",
+                "environment_id": "console-environment",
+                "states": {},
+                "contexes": {},
+                "history": [],
+                "current": null,
+                "meta": null,
+                "timestamp": Date.now(),
+                "data": {},
+                "created_at": Date.now(),
+                "updated_at": Date.now(),
+                "session_start": Date.now(),
+                "session_id": "test~from~console",
+                "id": "test~from~console"
+            };
+        }
+    }
+    resetSession() {
+        const jsonPath = `${os.homedir()}/.katasession`;
+        if (fs.existsSync(jsonPath)) {
+            fs.unlinkSync(jsonPath);
+            return true;
+        }
     }
 }
 exports.default = Bot;
