@@ -9,11 +9,7 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const merapi_1 = require("merapi");
-const colors = require("colors");
 const inquirer = require("inquirer");
-const repl = require("repl");
-const util = require("util");
-const deasync = require("deasync");
 class User extends merapi_1.Component {
     constructor(helper, api) {
         super();
@@ -86,7 +82,7 @@ class User extends merapi_1.Component {
                     else {
                         const token = result.data.id;
                         this.helper.setProp("first_login", { type: "user", username: user, id: result.data.userId });
-                        this.setToken({ name: user, type: "user" }, token);
+                        this.setToken({ name: user, type: "user", namespace: "platform" }, token);
                         console.log(`Logged in as ${user}`);
                     }
                 }
@@ -138,7 +134,9 @@ class User extends merapi_1.Component {
                         const result = yield this.helper.toPromise(this.api.authApi, this.api.authApi.tokensPost, { type: "team", teamId: teams[0].teamId });
                         const token = result.data.id;
                         this.setToken({ name, type: "team" }, token);
-                        console.log(`Switched to team ${name}`);
+                        this.helper.setProp("current_login", name);
+                        this.helper.setProp("current_user_type", "team");
+                        console.log(`Switched to team: ${name}`);
                     }
                     else {
                         console.log("Unable to switch to Team : Invalid team");
@@ -254,6 +252,7 @@ class User extends merapi_1.Component {
     setToken(userInfo, token) {
         this.helper.setProp("current_login", userInfo.name);
         this.helper.setProp("current_user_type", userInfo.type);
+        this.helper.setProp("namespace", userInfo.namespace);
         const tokenProp = (this.helper.getProp("token") || {});
         tokenProp[userInfo.name] = token;
         this.helper.setProp("token", tokenProp);
